jira-manager/requirements.txt
jira==3.5.2
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
ollama==0.1.6 